+++ data/Ultoi.txt
@@ -1,2 +1,1 @@
-event This is an event /at 2022-05-06 1234
-deadline This is a deadline /by 2022-02-09 2359
+todo aba
