--- lib/javafx.properties
@@ -1,3 +1,0 @@
-javafx.version=17.0.2-ea
-javafx.runtime.version=17.0.2-ea+3
-javafx.runtime.build=3
