--- src/main/java/ultoi/util/MainWindow.java
@@ -1,62 +1,0 @@
-package ultoi.util;
-
-import javafx.fxml.FXML;
-import javafx.scene.control.Button;
-import javafx.scene.control.ScrollPane;
-import javafx.scene.control.TextField;
-import javafx.scene.image.Image;
-import javafx.scene.layout.AnchorPane;
-import javafx.scene.layout.VBox;
-
-/**
- * Controller for MainWindow. Provides the layout for the other controls.
- */
-public class MainWindow extends AnchorPane {
-    @FXML
-    private ScrollPane scrollPane;
-    @FXML
-    private VBox dialogContainer;
-    @FXML
-    private TextField userInput;
-    @FXML
-    private Button sendButton;
-
-    private Ultoi ultoi;
-
-    private Image userImage = new Image(this.getClass().getResourceAsStream("/images/DaUser.png"));
-    private Image ultoiImage = new Image(this.getClass().getResourceAsStream("/images/DaUltoi.png"));
-
-    /**
-     * Initializes a main window.
-     */
-    @FXML
-    public void initialize() {
-        scrollPane.vvalueProperty().bind(dialogContainer.heightProperty());
-    }
-
-    /**
-     * Sets the driver for GUI.
-     *
-     * @param ultoi Driver for GUI.
-     */
-    public void setDriver(Ultoi ultoi) {
-        this.ultoi = ultoi;
-    }
-
-    /**
-     * Creates two dialog boxes, one echoing user input and the other containing Duke's reply and then appends them to
-     * the dialog container. Clears the user input after processing.
-     */
-    @FXML
-    private void handleUserInput() {
-        String userText = userInput.getText();
-        String ultoiText = ultoi.getResponse(userInput.getText());
-        dialogContainer.getChildren().addAll(
-                DialogBox.getUserDialog(userText, userImage),
-                (!ultoiText.contains(UltoiException.EXCEPTION_FACE))
-                        ? DialogBox.getUltoiDialog(ultoiText, ultoiImage)
-                        : DialogBox.getUltoiErrorDialog(ultoiText, ultoiImage)
-        );
-        userInput.clear();
-    }
-}
